+++ config/header-actions.php
@@ -1,0 +1,11 @@
+<?php
+
+return [
+    'primary_count' => 1,
+    'more' => [
+        'label' => 'More',
+        'icon' => 'heroicon-m-ellipsis-horizontal',
+        'color' => 'gray',
+        'hidden_label' => false,
+    ],
+];