+++ src/Commands/FilamentHeaderActionsCommand.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace VendorName\Skeleton\Commands;
+namespace Harvirsidhu\FilamentHeaderActions\Commands;
 
 use Illuminate\Console\Command;
 
-class SkeletonCommand extends Command
+class FilamentHeaderActionsCommand extends Command
 {
-    public $signature = 'skeleton';
+    public $signature = 'filament-header-actions';
 
     public $description = 'My command';