+++ src/FilamentHeaderActions.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Harvirsidhu\FilamentHeaderActions;
+
+use Harvirsidhu\FilamentHeaderActions\Actions\HeaderActionsComposer;
+
+class FilamentHeaderActions
+{
+    /**
+     * @param  array<mixed>  $actions
+     */
+    public function compose(array $actions): HeaderActionsComposer
+    {
+        return HeaderActionsComposer::make($actions);
+    }
+}