+++ src/Testing/TestsFilamentHeaderActions.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace VendorName\Skeleton\Testing;
+namespace Harvirsidhu\FilamentHeaderActions\Testing;
 
 use Livewire\Features\SupportTesting\Testable;
 
 /**
  * @mixin Testable
  */
-class TestsSkeleton
+class TestsFilamentHeaderActions
 {
     //
 }